Refactor profile.js
